backend/requirements.txt
aiobotocore==0.12.0
aiohttp==3.8.1
aioitertools==0.10.0
aiosignal==1.2.0
aiosqlite==0.17.0
alembic==1.4.1
appnope==0.1.2
argon2-cffi==21.1.0
async-timeout==4.0.2
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
bleach==4.1.0
botocore==1.15.15
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==7.1.2
cloudpickle==2.0.0
cryptography==36.0.2
cycler==0.11.0
databases==0.5.5
databricks-cli==0.16.2
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
docker==5.0.3
docutils==0.15.2
docx==0.2.4
docx2txt==0.8
entrypoints==0.3
fastapi==0.65.2
Flask==2.0.2
fonttools==4.28.1
frozenlist==1.3.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-core==2.2.3
google-cloud-documentai==1.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.45.0
grpcio-status==1.45.0
gunicorn==20.1.0
h11==0.9.0
h2o==3.34.0.3
httptools==0.1.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
llvmlite==0.37.0
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.21.0
multidict==6.0.2
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.7
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pdfminer==20191125
pdfminer.six==20220319
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==0.13.1
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.6
prompt-toolkit==3.0.22
proto-plus==1.20.3
protobuf==3.19.4
ptyprocess==0.7.0
py==1.11.0
py4j==0.10.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
Pygments==2.10.0
pyparsing==3.0.6
pyrsistent==0.18.0
pyspark==3.2.0
pytesseract==0.3.9
pytest==6.2.4
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.12.0
python-editor==1.0.4
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
querystring-parser==1.2.4
redis==4.0.2
regex==2022.3.15
requests==2.26.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
setuptools-scm==6.3.2
shap==0.40.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
starlette==0.14.2
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
urllib3==1.25.11
uvicorn==0.11.7
uvloop==0.16.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==8.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
xgboost==1.5.0
yarl==1.7.2
zipp==3.6.0
